+++ packages/gatsby-source-shopify/index.js
@@ -1,1 +1,10 @@
-// noop
+"use strict";
+
+exports.__esModule = true;
+
+var _constants = require("./constants");
+
+Object.keys(_constants).forEach(function (key) {
+  if (key === "default" || key === "__esModule") return;
+  exports[key] = _constants[key];
+});
