+++ packages/gatsby-source-shopify/package.json
@@ -44,5 +44,6 @@
   },
   "engines": {
     "node": ">=8.0.0"
-  }
+  },
+  "main": "index.js"
 }
