+++ packages/gatsby-source-shopify/src/index.js
@@ -1,0 +1,1 @@
+export * from "./constants"
